feat: add useEscapeKey
